Support for BitmapEffects
