+++ BuildFileTypeDialog.cs
@@ -124,7 +124,7 @@ namespace PaintDotNet.Effects
         {
             UpdateAllValues();
 
-            string sourceCode = ScriptWriter.FullFileTypeSourceCode(this.userCode, this.fileName, this.Author, this.Major, this.Minor, this.URL, this.Description, this.LoadExt, this.SaveExt, this.Layers, this.PluginName);
+            string sourceCode = FileTypeWriter.FullFileTypeSourceCode(this.userCode, this.fileName, this.Author, this.Major, this.Minor, this.URL, this.Description, this.LoadExt, this.SaveExt, this.Layers, this.PluginName);
             using (ViewSrc VSW = new ViewSrc("Full Source Code", sourceCode, true))
             {
                 VSW.ShowDialog();
@@ -149,7 +149,7 @@ namespace PaintDotNet.Effects
 
                 if (fbd.ShowDialog() == DialogResult.OK && !string.IsNullOrWhiteSpace(fbd.SelectedPath))
                 {
-                    string sourceCode = ScriptWriter.FullFileTypeSourceCode(this.userCode, this.fileName, this.Author, this.Major, this.Minor, this.URL, this.Description, this.LoadExt, this.SaveExt, this.Layers, this.PluginName);
+                    string sourceCode = FileTypeWriter.FullFileTypeSourceCode(this.userCode, this.fileName, this.Author, this.Major, this.Minor, this.URL, this.Description, this.LoadExt, this.SaveExt, this.Layers, this.PluginName);
                     string slnFilePath = Solution.Generate(fbd.SelectedPath, this.fileName, sourceCode, string.Empty, string.Empty);
 
                     if (slnFilePath != null)
