+++ CodeLab.csproj
@@ -207,6 +207,8 @@
     <EmbeddedResource Include="Icons\PlainText.png" />
     <EmbeddedResource Include="Icons\Book.png" />
     <EmbeddedResource Include="Icons\Docs.png" />
+    <EmbeddedResource Include="Icons\ClassicEffect.png" />
+    <EmbeddedResource Include="Icons\BitmapEffect.png" />
   </ItemGroup>
   <ItemGroup>
     <EmbeddedResource Include="Icons\CodeLab.32.png" />
@@ -254,6 +256,8 @@
     <EmbeddedResource Include="Icons\PlainText.32.png" />
     <EmbeddedResource Include="Icons\Book.32.png" />
     <EmbeddedResource Include="Icons\Docs.32.png" />
+    <EmbeddedResource Include="Icons\ClassicEffect.32.png" />
+    <EmbeddedResource Include="Icons\BitmapEffect.32.png" />
     <EmbeddedResource Include="Resources\Shape0.png" />
     <EmbeddedResource Include="Resources\Shape1.png" />
     <EmbeddedResource Include="Resources\Shape2.png" />
@@ -348,6 +352,6 @@
     <EmbeddedResource Include="Icons\WRKDST.png" />
   </ItemGroup>
   <Target Name="PostBuild" AfterTargets="PostBuildEvent">
-    <Exec Command="if $(ConfigurationName)==FastDebug goto end&#xD;&#xA;&#xD;&#xA;if not exist &quot;$(TargetDir)AllAssemblies\&quot; mkdir &quot;$(TargetDir)AllAssemblies\&quot;&#xD;&#xA;if not exist &quot;$(TargetDir)AllAssemblies\Merged\&quot; mkdir &quot;$(TargetDir)AllAssemblies\Merged&quot;&#xD;&#xA;&#xD;&#xA;set latestDotNet=NotFound&#xD;&#xA;&#xD;&#xA;for /F %25%25d in ('dir /a:d /b /o:-d &quot;C:\Program Files\dotnet\packs\Microsoft.NETCore.App.Ref\7.0.*&quot;') do (&#xD;&#xA;  set latestDotNet=%25%25d&#xD;&#xA;  goto loopExit&#xD;&#xA;)&#xD;&#xA;&#xD;&#xA;:loopExit&#xD;&#xA;&#xD;&#xA;copy $(ILRepack) &quot;$(TargetDir)AllAssemblies&quot;&#xD;&#xA;robocopy $(TargetDir) &quot;$(TargetDir)AllAssemblies&quot;&#xD;&#xA;robocopy &quot;C:\Program Files\dotnet\packs\Microsoft.NETCore.App.Ref\%25latestDotNet%25\ref\net7.0&quot; &quot;$(TargetDir)AllAssemblies&quot;&#xD;&#xA;robocopy &quot;C:\Program Files\dotnet\packs\Microsoft.WindowsDesktop.App.Ref\%25latestDotNet%25\ref\net7.0&quot; &quot;$(TargetDir)AllAssemblies&quot;&#xD;&#xA;&#xD;&#xA;set deps=ScintillaNET.dll Microsoft.CodeAnalysis.CSharp.dll Microsoft.CodeAnalysis.dll PluralizeService.Core.dll PlatformSpellCheck.dll&#xD;&#xA;&#xD;&#xA;cd $(TargetDir)AllAssemblies&#xD;&#xA;&#xD;&#xA;if $(ConfigurationName)==Release (&#xD;&#xA;    ILRepack.exe /ndebug /out:Merged\$(TargetFileName) $(TargetFileName) %25deps%25&#xD;&#xA;    cmd /c explorer &quot;$(TargetDir)AllAssemblies\Merged&quot;&#xD;&#xA;) else if $(ConfigurationName)==Debug (&#xD;&#xA;    ILRepack.exe /out:Merged\$(TargetFileName) $(TargetFileName) %25deps%25&#xD;&#xA;    copy &quot;$(TargetDir)AllAssemblies\Merged\$(TargetFileName)&quot; &quot;C:\Program Files\Paint.NET\Effects\&quot;&#xD;&#xA;)&#xD;&#xA;&#xD;&#xA;:end&#xD;&#xA;exit 0" />
+    <Exec Command="if $(ConfigurationName)==FastDebug goto end&#xD;&#xA;&#xD;&#xA;if not exist &quot;$(TargetDir)AllAssemblies\&quot; mkdir &quot;$(TargetDir)AllAssemblies\&quot;&#xD;&#xA;if not exist &quot;$(TargetDir)AllAssemblies\Merged\&quot; mkdir &quot;$(TargetDir)AllAssemblies\Merged&quot;&#xD;&#xA;&#xD;&#xA;set latestDotNet=NotFound&#xD;&#xA;&#xD;&#xA;for /F %25%25d in ('dir /a:d /b /o:-d &quot;C:\Program Files\dotnet\packs\Microsoft.NETCore.App.Ref\7.0.*&quot;') do (&#xD;&#xA;  set latestDotNet=%25%25d&#xD;&#xA;  goto loopExit&#xD;&#xA;)&#xD;&#xA;&#xD;&#xA;:loopExit&#xD;&#xA;&#xD;&#xA;copy $(ILRepack) &quot;$(TargetDir)AllAssemblies&quot;&#xD;&#xA;robocopy $(TargetDir) &quot;$(TargetDir)AllAssemblies&quot;&#xD;&#xA;robocopy &quot;C:\Program Files\dotnet\packs\Microsoft.NETCore.App.Ref\%25latestDotNet%25\ref\net7.0&quot; &quot;$(TargetDir)AllAssemblies&quot;&#xD;&#xA;robocopy &quot;C:\Program Files\dotnet\packs\Microsoft.WindowsDesktop.App.Ref\%25latestDotNet%25\ref\net7.0&quot; &quot;$(TargetDir)AllAssemblies&quot;&#xD;&#xA;&#xD;&#xA;set deps=ScintillaNET.dll Microsoft.CodeAnalysis.CSharp.dll Microsoft.CodeAnalysis.dll PluralizeService.Core.dll PlatformSpellCheck.dll&#xD;&#xA;&#xD;&#xA;cd $(TargetDir)AllAssemblies&#xD;&#xA;&#xD;&#xA;if $(ConfigurationName)==Release (&#xD;&#xA;    ILRepack.exe /ndebug /out:Merged\$(TargetFileName) $(TargetFileName) %25deps%25&#xD;&#xA;    cmd /c explorer &quot;$(TargetDir)AllAssemblies\Merged&quot;&#xD;&#xA;) else if $(ConfigurationName)==Debug (&#xD;&#xA;    copy &quot;$(TargetPath)&quot; &quot;C:\Program Files\paint.net\Effects\&quot;&#xD;&#xA;)&#xD;&#xA;&#xD;&#xA;:end&#xD;&#xA;exit 0" />
   </Target>
 </Project>
