+++ CodeLabConfigDialog.cs
@@ -18,11 +18,13 @@ using PaintDotNet.AppModel;
 using ScintillaNET;
 using System;
 using System.Collections.Generic;
+using System.Diagnostics;
 using System.Drawing;
 using System.IO;
 using System.Linq;
 using System.Text.RegularExpressions;
 using System.Threading.Tasks;
+using System.Windows.Controls.Primitives;
 using System.Windows.Forms;
 
 namespace PaintDotNet.Effects
@@ -79,6 +81,18 @@ namespace PaintDotNet.Effects
             opacity90MenuItem.Checked = false;
             opacity100MenuItem.Checked = true;
             transparencyToolStripMenuItem.Enabled = EnableOpacity;
+
+#if !RELEASE
+            ToolStripMenuItem[] items = Enum.GetValues<ProjectType>()
+                .Select(x => new ToolStripMenuItem(x.ToString(), null, (sender, e) => tabStrip1.NewTab(FileName, FullScriptPath, Enum.Parse<ProjectType>(((ToolStripMenuItem)sender).Text))))
+                .ToArray();
+
+            ToolStripDropDownButton newDocTypeChooser = new ToolStripDropDownButton();
+            newDocTypeChooser.DropDownItems.AddRange(items);
+            toolStrip1.Items.Insert(1, newDocTypeChooser);
+
+            //NewButton.DropDownItems.AddRange(items);
+#endif
         }
 
         protected override void OnLoading()
@@ -179,7 +193,7 @@ namespace PaintDotNet.Effects
         {
             FileName = token.ScriptName;
             FullScriptPath = token.ScriptPath;
-            if (token.ProjectType != ProjectType.Effect)
+            if (token.ProjectType != ProjectType.ClassicEffect)
             {
                 tabStrip1.SelectedTabIsDirty = false;
                 tabStrip1.NewTab(FileName, FullScriptPath, token.ProjectType);
@@ -203,7 +217,7 @@ namespace PaintDotNet.Effects
         {
             return new CodeLabConfigToken
             {
-                UserCode = ScriptWriter.DefaultCode,
+                UserCode = ClassicEffectWriter.DefaultCode,
                 UserScriptObject = null,
                 ScriptName = "Untitled",
                 ScriptPath = "",
@@ -227,8 +241,14 @@ namespace PaintDotNet.Effects
 
             switch (projType)
             {
-                case ProjectType.Effect:
-                    ScriptBuilder.Build<Effect>(txtCode.Text, OutputTextBox.Visible);
+                case ProjectType.ClassicEffect:
+                    ScriptBuilder.BuildClassicEffect(txtCode.Text, OutputTextBox.Visible);
+                    DisplayErrors();
+                    txtCode.UpdateSyntaxHighlighting();
+                    UpdateTokenFromDialog();
+                    break;
+                case ProjectType.BitmapEffect:
+                    ScriptBuilder.BuildBitmapEffect(txtCode.Text, OutputTextBox.Visible);
                     DisplayErrors();
                     txtCode.UpdateSyntaxHighlighting();
                     UpdateTokenFromDialog();
@@ -262,8 +282,11 @@ namespace PaintDotNet.Effects
             {
                 switch (projType)
                 {
-                    case ProjectType.Effect:
-                        ScriptBuilder.Build<Effect>(code, debug);
+                    case ProjectType.ClassicEffect:
+                        ScriptBuilder.BuildClassicEffect(code, debug);
+                        break;
+                    case ProjectType.BitmapEffect:
+                        ScriptBuilder.BuildBitmapEffect(code, debug);
                         break;
                     case ProjectType.FileType:
                         ScriptBuilder.BuildFileType(code, debug);
@@ -280,7 +303,9 @@ namespace PaintDotNet.Effects
 
             switch (projType)
             {
-                case ProjectType.Effect:
+                case ProjectType.ClassicEffect:
+                case ProjectType.GpuEffect:
+                case ProjectType.BitmapEffect:
                     txtCode.UpdateSyntaxHighlighting();
                     UpdateTokenFromDialog();
                     break;
@@ -293,7 +318,7 @@ namespace PaintDotNet.Effects
             tmrCompile.Enabled = true;
         }
 
-        private void RunEffectWithDialog()
+        private void RunEffectWithDialog(ProjectType projectType)
         {
             if (errorList.HasErrors)
             {
@@ -301,7 +326,18 @@ namespace PaintDotNet.Effects
                 return;
             }
 
-            if (!ScriptBuilder.BuildFullPreview(txtCode.Text))
+            bool built = false;
+            switch (projectType)
+            {
+                case ProjectType.ClassicEffect:
+                    built = ScriptBuilder.BuildClassicEffectFullPreview(txtCode.Text);
+                    break;
+                case ProjectType.BitmapEffect:
+                    built = ScriptBuilder.BuildBitmapEffectFullPreview(txtCode.Text);
+                    break;
+            }
+
+            if (!built)
             {
                 FlexibleMessageBox.Show("Something went wrong, and the Preview can't be run.", "Preview Error", MessageBoxButtons.OK, MessageBoxIcon.Error);
                 DisplayErrors();
@@ -412,7 +448,8 @@ namespace PaintDotNet.Effects
 
             switch (tabStrip1.SelectedTabProjType)
             {
-                case ProjectType.Effect:
+                case ProjectType.ClassicEffect:
+                case ProjectType.BitmapEffect:
                 case ProjectType.FileType:
                     if (ScriptBuilder.Errors.Count == 0)
                     {
@@ -491,7 +528,11 @@ namespace PaintDotNet.Effects
                 case ".cs":
                     if (Regex.IsMatch(fileContents, @"void\s+Render\s*\(\s*Surface\s+dst\s*,\s*Surface\s+src\s*,\s*Rectangle\s+rect\s*\)\s*{(.|\s)*}", RegexOptions.Singleline))
                     {
-                        projType = ProjectType.Effect;
+                        projType = ProjectType.ClassicEffect;
+                    }
+                    else if (Regex.IsMatch(fileContents, @"protected\s+override\s+void\s+OnRender\s*\(\s*IBitmapEffectOutput\s+output\s*\)\s*{(.|\s)*}", RegexOptions.Singleline))
+                    {
+                        projType = ProjectType.BitmapEffect;
                     }
                     else if (Regex.IsMatch(fileContents, @"void\s+SaveImage\s*\(\s*Document\s+input\s*,\s*Stream\s+output\s*,\s*PropertyBasedSaveConfigToken\s+token\s*,\s*Surface\s+scratchSurface\s*,\s*ProgressEventHandler\s+progressCallback\s*\)\s*{(.|\s)*}", RegexOptions.Singleline))
                     {
@@ -585,7 +626,7 @@ namespace PaintDotNet.Effects
                     description = "XAML Shape";
                     isCSharp = false;
                     break;
-                case ProjectType.Effect:
+                case ProjectType.ClassicEffect:
                 case ProjectType.FileType:
                 default:
                     fileExt = ".cs";
@@ -943,13 +984,13 @@ namespace PaintDotNet.Effects
                 FileName = "Untitled";
                 FullScriptPath = "";
 
-                if (tabStrip1.SelectedTabIsInitial && tabStrip1.SelectedTabProjType == ProjectType.Effect && txtCode.IsVirgin)
+                if (tabStrip1.SelectedTabIsInitial && tabStrip1.SelectedTabProjType == ProjectType.ClassicEffect && txtCode.IsVirgin)
                 {
                     UpdateTabProperties();
                 }
                 else
                 {
-                    tabStrip1.NewTab(FileName, FullScriptPath, ProjectType.Effect);
+                    tabStrip1.NewTab(FileName, FullScriptPath, ProjectType.ClassicEffect);
                 }
 
                 txtCode.Text = fn.CodeTemplate;
@@ -1167,37 +1208,63 @@ namespace PaintDotNet.Effects
 
             bool buildSucceeded = false;
 #if FASTDEBUG
-                    const bool isClassic = false;
+            const bool canBuildSln = false;
 #else
-            bool isClassic = this.Services.GetService<IAppInfoService>().InstallType == AppInstallType.Classic;
+            bool canBuildSln = this.Services.GetService<IAppInfoService>().InstallType == AppInstallType.Classic;
 #endif
 
             switch (tabStrip1.SelectedTabProjType)
             {
-                case ProjectType.Effect:
-                    BuildForm buildForm = new BuildForm(fileName, txtCode.Text, FullScriptPath, isClassic);
-                    if (buildForm.ShowDialog() != DialogResult.OK)
+                case ProjectType.ClassicEffect:
+                    using (BuildForm buildForm = new BuildForm(fileName, txtCode.Text, FullScriptPath, ProjectType.ClassicEffect, canBuildSln))
                     {
-                        return;
+                        if (buildForm.ShowDialog() != DialogResult.OK)
+                        {
+                            return;
+                        }
+
+                        string sourceCode = ClassicEffectWriter.FullSourceCode(
+                            txtCode.Text, Path.GetFileNameWithoutExtension(FullScriptPath), buildForm.isAdjustment,
+                            buildForm.SubMenu, buildForm.MenuItemName, buildForm.IconPath, buildForm.URL, buildForm.RenderingFlags,
+                            buildForm.RenderingSchedule, buildForm.Author, buildForm.MajorVer, buildForm.MinorVer,
+                            buildForm.Description, buildForm.KeyWords, buildForm.WindowTitle, buildForm.HelpType, buildForm.HelpStr);
+
+                        buildSucceeded = ScriptBuilder.BuildEffectDll(sourceCode, FullScriptPath, buildForm.IconPath, buildForm.HelpType);
                     }
 
-                    buildSucceeded = ScriptBuilder.BuildEffectDll(
-                        txtCode.Text, FullScriptPath, buildForm.SubMenu, buildForm.MenuItemName, buildForm.IconPath, buildForm.Author, buildForm.MajorVer, buildForm.MinorVer, buildForm.URL,
-                        buildForm.WindowTitle, buildForm.isAdjustment, buildForm.Description, buildForm.KeyWords, buildForm.EffectFlags, buildForm.RenderingSchedule, buildForm.HelpType, buildForm.HelpStr);
+                    break;
+                case ProjectType.BitmapEffect:
+                    using (BuildForm buildForm = new BuildForm(fileName, txtCode.Text, FullScriptPath, ProjectType.BitmapEffect, canBuildSln))
+                    {
+                        if (buildForm.ShowDialog() != DialogResult.OK)
+                        {
+                            return;
+                        }
 
-                    buildForm.Dispose();
+                        string sourceCode = BitmapEffectWriter.FullSourceCode(
+                            txtCode.Text, Path.GetFileNameWithoutExtension(FullScriptPath), buildForm.isAdjustment,
+                            buildForm.SubMenu, buildForm.MenuItemName, buildForm.IconPath, buildForm.URL, buildForm.RenderingFlags,
+                            buildForm.RenderingSchedule, buildForm.Author, buildForm.MajorVer, buildForm.MinorVer,
+                            buildForm.Description, buildForm.KeyWords, buildForm.WindowTitle, buildForm.HelpType, buildForm.HelpStr);
+
+                        buildSucceeded = ScriptBuilder.BuildEffectDll(sourceCode, FullScriptPath, buildForm.IconPath, buildForm.HelpType);
+                    }
+
                     break;
                 case ProjectType.FileType:
-                    BuildFileTypeDialog buildFileTypeDialog = new BuildFileTypeDialog(fileName, txtCode.Text, isClassic);
-                    if (buildFileTypeDialog.ShowDialog() != DialogResult.OK)
+                    using (BuildFileTypeDialog buildFileTypeDialog = new BuildFileTypeDialog(fileName, txtCode.Text, canBuildSln))
                     {
-                        return;
+                        if (buildFileTypeDialog.ShowDialog() != DialogResult.OK)
+                        {
+                            return;
+                        }
+
+                        buildSucceeded = ScriptBuilder.BuildFileTypeDll(
+                            txtCode.Text, FullScriptPath, buildFileTypeDialog.Author, buildFileTypeDialog.Major, buildFileTypeDialog.Minor,
+                            buildFileTypeDialog.URL, buildFileTypeDialog.Description, buildFileTypeDialog.LoadExt, buildFileTypeDialog.SaveExt,
+                            buildFileTypeDialog.Layers, buildFileTypeDialog.PluginName);
                     }
 
-                    buildSucceeded = ScriptBuilder.BuildFileTypeDll(txtCode.Text, FullScriptPath, buildFileTypeDialog.Author, buildFileTypeDialog.Major, buildFileTypeDialog.Minor,
-                        buildFileTypeDialog.URL, buildFileTypeDialog.Description, buildFileTypeDialog.LoadExt, buildFileTypeDialog.SaveExt, buildFileTypeDialog.Layers, buildFileTypeDialog.PluginName);
-
-                    buildFileTypeDialog.Dispose();
                     break;
             }
 
@@ -1363,10 +1430,12 @@ namespace PaintDotNet.Effects
             tmrCompile.Enabled = false;
             Build();
 
-            switch (this.tabStrip1.SelectedTabProjType)
+            ProjectType projectType = this.tabStrip1.SelectedTabProjType;
+            switch (projectType)
             {
-                case ProjectType.Effect:
-                    RunEffectWithDialog();
+                case ProjectType.ClassicEffect:
+                case ProjectType.BitmapEffect:
+                    RunEffectWithDialog(projectType);
                     break;
                 case ProjectType.FileType:
                     RunFileTypeWithDialog();
@@ -1383,7 +1452,7 @@ namespace PaintDotNet.Effects
         {
             ProjectType projectType = tabStrip1.SelectedTabProjType;
             bool notRef = projectType != ProjectType.Reference;
-            bool cSharp = projectType == ProjectType.Effect || projectType == ProjectType.FileType;
+            bool cSharp = projectType.IsCSharp();
 
             saveToolStripMenuItem.Enabled = notRef;
             saveAsToolStripMenuItem.Enabled = notRef;
@@ -1454,7 +1523,7 @@ namespace PaintDotNet.Effects
         {
             bool hasText = txtCode.TextLength > 0;
             bool isTextSelected = hasText && txtCode.SelectedText.Length > 0;
-            bool cSharp = tabStrip1.SelectedTabProjType == ProjectType.Effect || tabStrip1.SelectedTabProjType == ProjectType.FileType;
+            bool cSharp = tabStrip1.SelectedTabProjType.IsCSharp();
 
             this.cutToolStripMenuItem1.Enabled = isTextSelected;
             this.copyToolStripMenuItem1.Enabled = isTextSelected;
@@ -1695,7 +1764,7 @@ namespace PaintDotNet.Effects
 
             bool hasText = txtCode.TextLength > 0;
             bool isTextSelected = hasText && txtCode.SelectedText.Length > 0;
-            bool cSharp = tabStrip1.SelectedTabProjType == ProjectType.Effect || tabStrip1.SelectedTabProjType == ProjectType.FileType;
+            bool cSharp = tabStrip1.SelectedTabProjType == ProjectType.ClassicEffect || tabStrip1.SelectedTabProjType == ProjectType.FileType;
 
             this.cutToolStripMenuItem.Enabled = isTextSelected;
             this.copyToolStripMenuItem.Enabled = isTextSelected;
@@ -2037,8 +2106,7 @@ namespace PaintDotNet.Effects
             UpdateToolBarButtons();
             DisableButtonsForRef(projectType == ProjectType.Reference);
 
-            if (projectType == ProjectType.Effect ||
-                projectType == ProjectType.FileType)
+            if (projectType.IsCSharp())
             {
                 EnableCSharpButtons(true);
                 BuildAsync();
@@ -2060,7 +2128,7 @@ namespace PaintDotNet.Effects
             UpdateWindowTitle();
             UpdateToolBarButtons();
             DisableButtonsForRef(projectType == ProjectType.Reference);
-            EnableCSharpButtons(projectType == ProjectType.Effect || projectType == ProjectType.FileType);
+            EnableCSharpButtons(projectType.IsCSharp());
         }
 
         private void EnableCSharpButtons(bool enable)
@@ -2155,15 +2223,5 @@ namespace PaintDotNet.Effects
         }
         #endregion
 
-    }
-
-    public enum ProjectType
-    {
-        None,
-        Effect,
-        EffectGpu,
-        FileType,
-        Reference,
-        Shape
     }
 }
